pyproject.toml
[project]
name = "tool-timing-channel"
version = "0.1.0"
description = "Covert channel taxonomy for AI agent tool calls"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.60.0",
    "httpx>=0.28.0",
    "numpy>=2.0.0",
    "matplotlib>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
